Changed I2CEEPROMRead and added a little more I2C delay
Trying to get to stop locking up randomly on eeprom I2C read, hasn't
helped, but probably good to leave in.

+++ Source/CommCtrlrArduino/LongMsg.h
@@ -31,8 +31,9 @@ private:
         unsigned int _StartAddr;
         unsigned int _EndAddr;
 		
-        byte I2CeePROMRead(byte device, unsigned int addr);
+        int I2CeePROMRead(byte device, unsigned int addr);
 };
 
 #endif
+
